LayoutTests/imported/w3c/web-platform-tests/css/css-conditional/container-queries/container-nested-expected.txt


FAIL Implicit assert_equals: expected "true" but got ""
PASS Implicit, outer failing
PASS Implicit, inner failing
FAIL Outer named, inner named assert_equals: expected "true" but got ""
FAIL Outer named, inner named (reverse) assert_equals: expected "true" but got ""
PASS Failing outer name
PASS Failing inner name
FAIL Outer named, inner implicit assert_equals: expected "true" but got ""
FAIL Inner named, outer implicit assert_equals: expected "true" but got ""
FAIL Inner named, outer implicit (reverse) assert_equals: expected "true" but got ""
FAIL Three levels assert_equals: expected "true" but got ""
PASS Three levels, middle fail
FAIL Named inner invalidation assert_equals: expected "true" but got ""
FAIL Implicit outer invalidation assert_equals: expected "true" but got ""
